+++ demisto_sdk/yaml_tools/unifier.py
@@ -4,24 +4,36 @@ import glob
 import yaml
 import base64
 import re
+from typing import Tuple
 
 from demisto_sdk.common.constants import Errors
-from demisto_sdk.common.tools import get_yaml, server_version_compare, get_yml_paths_in_dir
+from demisto_sdk.common.tools import get_yaml, server_version_compare, get_yml_paths_in_dir, print_error, print_color, \
+    LOG_COLORS
 from demisto_sdk.common.constants import TYPE_TO_EXTENSION, INTEGRATIONS_DIR, DIR_TO_PREFIX, DEFAULT_IMAGE_PREFIX, \
     SCRIPTS_DIR, BETA_INTEGRATIONS_DIR
 
 
 class Unifier:
 
-    def __init__(self, package_path: str, dir_name=INTEGRATIONS_DIR, dest_path='', image_prefix=DEFAULT_IMAGE_PREFIX):
+    def __init__(self, indir: str, dir_name=INTEGRATIONS_DIR, outdir='',
+                 image_prefix=DEFAULT_IMAGE_PREFIX):
+
+        directory_name = ""
+        for optional_dir_name in DIR_TO_PREFIX:
+            if optional_dir_name in indir:
+                directory_name = optional_dir_name
+
+        if not directory_name:
+            print_error("You have failed to provide a legal file path, a legal file path "
+                        "should contain either Integrations or Scripts directories")
 
         self.image_prefix = image_prefix
-        self.package_path = package_path
+        self.package_path = indir
         if self.package_path[-1] != os.sep:
             self.package_path = os.path.join(self.package_path, '')
 
         self.dir_name = dir_name
-        self.dest_path = dest_path
+        self.dest_path = outdir
 
         self.is_ci = os.getenv('CI', False)
 
@@ -80,9 +92,6 @@ class Unifier:
 
     def merge_script_package_to_yml(self):
         """Merge the various components to create an output yml file
-
-        Returns:
-            output path, script path, image path
         """
         print("Merging package: {}".format(self.package_path))
         if self.package_path.endswith('/'):
@@ -125,7 +134,8 @@ class Unifier:
             yml_text, desc_path = self.insert_description_to_yml(yml_data, yml_text)
 
         output_map = self.write_yaml_with_docker(yml_text, yml_data, script_obj)
-        return list(output_map.keys()), yml_path, script_path, image_path, desc_path
+        unifier_outputs = list(output_map.keys()), yml_path, script_path, image_path, desc_path
+        print_color("Created unified yml: {}".format(unifier_outputs[0][0]), LOG_COLORS.GREEN)
 
     def insert_image_to_yml(self, yml_data, yml_text):
         image_data, found_img_path = self.get_data("*png")
@@ -184,11 +194,13 @@ class Unifier:
         """
 
         ignore_regex = (r'CommonServerPython\.py|CommonServerUserPython\.py|demistomock\.py|_test\.py'
-                        r'|conftest\.py|__init__\.py')
+                        r'|conftest\.py|__init__\.py|ApiModule\.py')
         if not self.package_path.endswith('/'):
             self.package_path += '/'
         if self.package_path.endswith('Scripts/CommonServerPython/'):
             return self.package_path + 'CommonServerPython.py'
+        if self.package_path.endswith('ApiModule/'):
+            return os.path.join(self.package_path, os.path.basename(os.path.normpath(self.package_path)) + '.py')
 
         script_path = list(filter(lambda x: not re.search(ignore_regex, x),
                                   glob.glob(os.path.join(self.package_path, '*' + script_type))))[0]
@@ -198,6 +210,12 @@ class Unifier:
         script_path = self.get_code_file(script_type)
         with io.open(script_path, mode='r', encoding='utf-8') as script_file:
             script_code = script_file.read()
+
+        # Check if the script imports an API module. If it does,
+        # the API module code will be pasted in place of the import.
+        module_import, module_name = self.check_api_module_imports(script_code)
+        if module_import:
+            script_code = self.insert_module_code(script_code, module_import, module_name)
 
         clean_code = self.clean_python_code(script_code)
 
@@ -247,6 +265,57 @@ class Unifier:
         return yml_path, code
 
     @staticmethod
+    def check_api_module_imports(script_code: str) -> Tuple[str, str]:
+        """
+        Checks integration code for API module imports
+        :param script_code: The integration code
+        :return: The import string and the imported module name
+        """
+
+        # General regex to find API module imports, for example: "from MicrosoftApiModule import *  # noqa: E402"
+        module_regex = r'from ([\w\d]+ApiModule) import \*(?:  # noqa: E402)?'
+
+        module_match = re.search(module_regex, script_code)
+        if module_match:
+            return module_match.group(), module_match.group(1)
+        else:
+            return '', ''
+
+    @staticmethod
+    def insert_module_code(script_code: str, module_import: str, module_name: str) -> str:
+        """
+        Inserts API module in place of an import to the module according to the module name
+        :param script_code: The integration code
+        :param module_import: The module import string to replace
+        :param module_name: The module name
+        :return: The integration script with the module code appended in place of the import
+        """
+
+        module_path = os.path.join('./Packs', 'ApiModules', 'Scripts', module_name, module_name + '.py')
+        module_code = Unifier._get_api_module_code(module_name, module_path)
+
+        module_code = '\n### GENERATED CODE ###\n# This code was inserted in place of an API module.{}\n' \
+            .format(module_code)
+
+        return script_code.replace(module_import, module_code)
+
+    @staticmethod
+    def _get_api_module_code(module_name, module_path):
+        """
+        Attempts to get the API module code from the ApiModules pack.
+        :param module_name: The API module name
+        :param module_path: The API module code file path
+        :return: The API module code
+        """
+        try:
+            with io.open(module_path, mode='r', encoding='utf-8') as script_file:
+                module_code = script_file.read()
+        except Exception as e:
+            raise ValueError('Could not retrieve the module [{}] code: {}'.format(module_name, str(e)))
+
+        return module_code
+
+    @staticmethod
     def clean_python_code(script_code, remove_print_future=True):
         script_code = script_code.replace("import demistomock as demisto", "")
         script_code = script_code.replace("from CommonServerPython import *", "")
@@ -255,10 +324,3 @@ class Unifier:
         if remove_print_future:  # docs generation requires to leave this
             script_code = script_code.replace("from __future__ import print_function", "")
         return script_code
-
-    @staticmethod
-    def add_sub_parser(subparsers):
-        parser = subparsers.add_parser('unify',
-                                       help='Unify code, image and description files to a single Demisto yaml file')
-        parser.add_argument("-i", "--indir", help="The path to the files to unify", required=True)
-        parser.add_argument("-o", "--outdir", help="The output dir to write the unified yml to", required=True)